Added missing dependencies used for relocation when creating shaded jar
The SDK has some dependencies that can conflict with Jersey 2.0.
To resolve this, an option to create a shaded JAR is provided:
https://github.com/qubole/qds-sdk-java#jersey-20

Creating a shaded JAR file includes all the dependencies of the SDK
in a single JAR while modifying the namespace of the included
dependencies to avoid conflict. However, not all dependencies are
relocated in the POM file; dependencies with the namespace
`com.fasterxml.jackson` are missed.

Adding the missing dependencies for relocation in the POM file.